+++ src/lib/pages/contract-list/components/ContractListTable.tsx
@@ -84,7 +84,7 @@ export const ContractListTable = ({
             <Th width="15%">Contract Address</Th>
             <Th width="25%">Contract Name</Th>
             <Th width="25%">Tags</Th>
-            <Th width="25%">Instantiator</Th>
+            <Th width="25%">Instantiated by</Th>
             <Th width="10%" />
           </Tr>
         </Thead>
@@ -125,7 +125,7 @@ export const ContractListTable = ({
                 <Td>
                   <TagsCell contractLocalInfo={item} />
                 </Td>
-                {/* Instantiator */}
+                {/* Instantiated by */}
                 <Td>
                   <ExplorerLink
                     value={item.instantiator}