+++ src/lib/services/contractService.ts
@@ -2,7 +2,7 @@ import type { UseQueryResult } from "@tanstack/react-query";
 import { useQuery } from "@tanstack/react-query";
 import { useCallback } from "react";
 
-import { indexerGraphClient } from "lib/data/graphql";
+import { useCelatoneApp } from "lib/app-provider";
 import {
   getInstantiatedListByUserQueryDocument,
   getInstantiatedCountByUserQueryDocument,
@@ -38,6 +38,7 @@ import {
   parseTxHash,
   parseTxHashOpt,
   snakeToCamel,
+  unwrap,
 } from "lib/utils";
 
 interface InstantiateDetail {
@@ -50,6 +51,7 @@ interface InstantiateDetail {
 export const useInstantiatedCountByUserQuery = (
   walletAddr: Option<HumanAddr>
 ): UseQueryResult<Option<number>> => {
+  const { indexerGraphClient } = useCelatoneApp();
   const queryFn = useCallback(async () => {
     if (!walletAddr) return undefined;
 
@@ -58,7 +60,7 @@ export const useInstantiatedCountByUserQuery = (
         walletAddr,
       })
       .then(({ contracts_aggregate }) => contracts_aggregate?.aggregate?.count);
-  }, [walletAddr]);
+  }, [indexerGraphClient, walletAddr]);
 
   // TODO: add query key later
   return useQuery(["instantiated_count_by_user", walletAddr], queryFn, {
@@ -70,6 +72,7 @@ export const useInstantiatedCountByUserQuery = (
 export const useInstantiatedListByUserQuery = (
   walletAddr: Option<HumanAddr>
 ): UseQueryResult<Option<ContractLocalInfo[]>> => {
+  const { indexerGraphClient } = useCelatoneApp();
   const queryFn = useCallback(async () => {
     if (!walletAddr) return undefined;
 
@@ -80,11 +83,11 @@ export const useInstantiatedListByUserQuery = (
       .then(({ contracts }) =>
         contracts.map<ContractLocalInfo>((contract) => ({
           contractAddress: contract.address as ContractAddr,
-          instantiator: walletAddr,
+          instantiator: unwrap(contract.accountByInitBy?.address),
           label: contract.label,
         }))
       );
-  }, [walletAddr]);
+  }, [indexerGraphClient, walletAddr]);
 
   // TODO: add query key later
   return useQuery(["instantiated_list_by_user", walletAddr], queryFn, {
@@ -96,6 +99,7 @@ export const useInstantiatedListByUserQuery = (
 export const useContractListByAdmin = (
   adminAddress: Option<ContractAddr | HumanAddr>
 ): UseQueryResult<Option<ContractLocalInfo[]>> => {
+  const { indexerGraphClient } = useCelatoneApp();
   const queryFn = useCallback(async () => {
     if (!adminAddress) return undefined;
 
@@ -110,7 +114,7 @@ export const useContractListByAdmin = (
           label: contract.label,
         }))
       );
-  }, [adminAddress]);
+  }, [adminAddress, indexerGraphClient]);
 
   return useQuery(["contract_list_by_admin", adminAddress], queryFn, {
     keepPreviousData: true,
@@ -121,6 +125,7 @@ export const useContractListByAdmin = (
 export const useInstantiateDetailByContractQuery = (
   contractAddress: ContractAddr
 ): UseQueryResult<InstantiateDetail> => {
+  const { indexerGraphClient } = useCelatoneApp();
   const queryFn = useCallback(async () => {
     return indexerGraphClient
       .request(getInstantiateDetailByContractQueryDocument, { contractAddress })
@@ -131,7 +136,7 @@ export const useInstantiateDetailByContractQuery = (
         initProposalTitle:
           contracts_by_pk?.contract_proposals.at(0)?.proposal.title,
       }));
-  }, [contractAddress]);
+  }, [contractAddress, indexerGraphClient]);
 
   return useQuery(
     ["instantiate_detail_by_contract", contractAddress],
@@ -145,6 +150,7 @@ export const useInstantiateDetailByContractQuery = (
 export const useAdminByContractAddresses = (
   contractAddresses: Option<ContractAddr[]>
 ): UseQueryResult<Option<Record<ContractAddr, string>>> => {
+  const { indexerGraphClient } = useCelatoneApp();
   const queryFn = useCallback(async () => {
     if (!contractAddresses) return undefined;
 
@@ -161,7 +167,7 @@ export const useAdminByContractAddresses = (
           {}
         )
       );
-  }, [contractAddresses]);
+  }, [contractAddresses, indexerGraphClient]);
 
   return useQuery(["admin_by_contracts", contractAddresses], queryFn, {
     keepPreviousData: true,
@@ -174,6 +180,7 @@ export const useExecuteTxsByContractAddress = (
   offset: number,
   pageSize: number
 ): UseQueryResult<Option<ExecuteTransaction[]>> => {
+  const { indexerGraphClient } = useCelatoneApp();
   const queryFn = useCallback(async () => {
     return indexerGraphClient
       .request(getExecuteTxsByContractAddress, {
@@ -193,7 +200,7 @@ export const useExecuteTxsByContractAddress = (
           success: transaction.transaction.success,
         }))
       );
-  }, [contractAddress, offset, pageSize]);
+  }, [contractAddress, offset, pageSize, indexerGraphClient]);
 
   return useQuery(
     [
@@ -213,6 +220,7 @@ export const useExecuteTxsByContractAddress = (
 export const useExecuteTxsCountByContractAddress = (
   contractAddress: ContractAddr
 ): UseQueryResult<Option<number>> => {
+  const { indexerGraphClient } = useCelatoneApp();
   const queryFn = useCallback(async () => {
     if (!contractAddress) return undefined;
 
@@ -224,7 +232,7 @@ export const useExecuteTxsCountByContractAddress = (
         ({ contract_transactions_aggregate }) =>
           contract_transactions_aggregate?.aggregate?.count
       );
-  }, [contractAddress]);
+  }, [contractAddress, indexerGraphClient]);
 
   return useQuery(
     ["execute_transactions_count_by_contract_addr", contractAddress],
@@ -243,6 +251,7 @@ export const useMigrationHistoriesByContractAddress = (
 ): UseQueryResult<
   Option<Omit<ContractMigrationHistory, "codeDescription">[]>
 > => {
+  const { indexerGraphClient } = useCelatoneApp();
   const queryFn = useCallback(async () => {
     return indexerGraphClient
       .request(getMigrationHistoriesByContractAddress, {
@@ -265,7 +274,7 @@ export const useMigrationHistoriesByContractAddress = (
           },
         }))
       );
-  }, [contractAddress, offset, pageSize]);
+  }, [contractAddress, offset, pageSize, indexerGraphClient]);
 
   return useQuery(
     ["migration_histories", contractAddress, offset, pageSize],
@@ -280,6 +289,7 @@ export const useMigrationHistoriesByContractAddress = (
 export const useMigrationHistoriesCountByContractAddress = (
   contractAddress: ContractAddr
 ): UseQueryResult<Option<number>> => {
+  const { indexerGraphClient } = useCelatoneApp();
   const queryFn = useCallback(async () => {
     return indexerGraphClient
       .request(getMigrationHistoriesCountByContractAddress, {
@@ -289,7 +299,7 @@ export const useMigrationHistoriesCountByContractAddress = (
         ({ contract_histories_aggregate }) =>
           contract_histories_aggregate.aggregate?.count
       );
-  }, [contractAddress]);
+  }, [contractAddress, indexerGraphClient]);
 
   return useQuery(["migration_histories_count", contractAddress], queryFn, {
     keepPreviousData: true,
@@ -302,6 +312,7 @@ export const useTxsByContractAddress = (
   offset: number,
   pageSize: number
 ): UseQueryResult<AllTransaction[]> => {
+  const { indexerGraphClient } = useCelatoneApp();
   const queryFn = useCallback(async () => {
     return indexerGraphClient
       .request(getTxsByContractAddress, {
@@ -331,7 +342,7 @@ export const useTxsByContractAddress = (
           isIbc: contractTx.transaction.is_ibc,
         }))
       );
-  }, [contractAddress, offset, pageSize]);
+  }, [contractAddress, offset, pageSize, indexerGraphClient]);
 
   return useQuery(
     ["transactions_by_contract_addr", contractAddress, offset, pageSize],
@@ -346,6 +357,7 @@ export const useTxsByContractAddress = (
 export const useTxsCountByContractAddress = (
   contractAddress: ContractAddr
 ): UseQueryResult<Option<number>> => {
+  const { indexerGraphClient } = useCelatoneApp();
   const queryFn = useCallback(async () => {
     if (!contractAddress) return undefined;
 
@@ -357,7 +369,7 @@ export const useTxsCountByContractAddress = (
         ({ contract_transactions_aggregate }) =>
           contract_transactions_aggregate.aggregate?.count
       );
-  }, [contractAddress]);
+  }, [contractAddress, indexerGraphClient]);
 
   return useQuery(
     ["transactions_count_by_contract_addr", contractAddress],
@@ -374,6 +386,7 @@ export const useRelatedProposalsByContractAddress = (
   offset: number,
   pageSize: number
 ): UseQueryResult<Option<ContractRelatedProposals[]>> => {
+  const { indexerGraphClient } = useCelatoneApp();
   const queryFn = useCallback(async () => {
     return indexerGraphClient
       .request(getRelatedProposalsByContractAddress, {
@@ -395,7 +408,7 @@ export const useRelatedProposalsByContractAddress = (
             | ContractAddr,
         }))
       );
-  }, [contractAddress, offset, pageSize]);
+  }, [contractAddress, offset, pageSize, indexerGraphClient]);
 
   return useQuery(
     ["related_proposals", contractAddress, offset, pageSize],
@@ -410,6 +423,7 @@ export const useRelatedProposalsByContractAddress = (
 export const useRelatedProposalsCountByContractAddress = (
   contractAddress: ContractAddr
 ): UseQueryResult<Option<number>> => {
+  const { indexerGraphClient } = useCelatoneApp();
   const queryFn = useCallback(async () => {
     return indexerGraphClient
       .request(getRelatedProposalsCountByContractAddress, {
@@ -419,7 +433,7 @@ export const useRelatedProposalsCountByContractAddress = (
         ({ contract_proposals_aggregate }) =>
           contract_proposals_aggregate.aggregate?.count
       );
-  }, [contractAddress]);
+  }, [contractAddress, indexerGraphClient]);
 
   return useQuery(["related_proposals_count", contractAddress], queryFn, {
     keepPreviousData: true,