+++ src/pages/public-project/index.tsx
@@ -1,0 +1,3 @@
+import { AllPublicProjectsPage } from "lib/pages/public-project";
+
+export default AllPublicProjectsPage;